Move time-related classes to utils.time package
Updated Timer, Scheduler, TimerTest, and Counter package declarations for better organization.
Added package-info.java to document the utils.time package.
