+++ util/src/main/java/org/akazukin/util/utils/Counter.java
@@ -1,4 +1,4 @@
-package org.akazukin.util.utils.timer;
+package org.akazukin.util.utils;
 
 import lombok.AccessLevel;
 import lombok.Getter;
