+++ util/src/main/java/org/akazukin/util/utils/time/Scheduler.java
@@ -1,4 +1,4 @@
-package org.akazukin.util.utils;
+package org.akazukin.util.utils.time;
 
 import lombok.AccessLevel;
 import lombok.SneakyThrows;
