+++ util/src/main/java/org/akazukin/util/utils/time/Timer.java
@@ -1,4 +1,4 @@
-package org.akazukin.util.utils.timer;
+package org.akazukin.util.utils.time;
 
 import lombok.AccessLevel;
 import lombok.experimental.FieldDefaults;
