+++ util/src/main/java/org/akazukin/util/utils/time/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Provides for time-related utilities
+ *
+ * @author currypan1229
+ * @since 1.0.0
+ */
+package org.akazukin.util.utils.time;
